Split out IDataSerializable interface.
