+++ clicache/src/IGeodeSerializable.hpp
@@ -62,48 +62,6 @@ namespace Apache
         String^ ToString( );
       };
 
-      public interface class IDataSerializable : public IGeodeSerializable
-      {
-      public:
-        /// <summary>
-        /// Serializes this object.
-        /// </summary>
-        /// <param name="output">
-        /// the DataOutput object to use for serializing the object
-        /// </param>
-        void ToData( DataOutput^ output );
-
-        //bool HasDelta();
-
-        /// <summary>
-        /// Deserialize this object, typical implementation should return
-        /// the 'this' pointer.
-        /// </summary>
-        /// <param name="input">
-        /// the DataInput stream to use for reading the object data
-        /// </param>
-        /// <returns>the deserialized object</returns>
-        void FromData( DataInput^ input );
-
-        /// <summary>
-        /// Returns the classId of the instance being serialized.
-        /// This is used by deserialization to determine what instance
-        /// type to create and deserialize into.
-        /// </summary>
-        /// <remarks>
-        /// The classId must be unique within an application suite
-        /// and in the range 0 to ((2^31)-1) both inclusive. An application can
-        /// thus define upto 2^31 custom <c>IGeodeSerializable</c> classes.
-        /// Returning a value greater than ((2^31)-1) may result in undefined
-        /// behaviour.
-        /// </remarks>
-        /// <returns>the classId</returns>
-        property System::UInt32 ClassId
-        {
-          System::UInt32 get( );
-        }
-      };
-
       public interface class IDataSerializablePrimitive : public IGeodeSerializable
       {
         void ToData( DataOutput^ output );
